Name changes and better structure.
