--- src/main/java/com/example/Graduation/Project/GraduationProjectApplication.java
@@ -1,13 +1,0 @@
-package com.example.Graduation.Project;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class GraduationProjectApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(GraduationProjectApplication.class, args);
-	}
-
-}
