--- src/main/java/com/example/Graduation/Project/entity/HealthQuestionReference.java
@@ -1,19 +1,0 @@
-package com.example.Graduation.Project.entity;
-
-import jakarta.persistence.*;
-import lombok.Getter;
-import lombok.Setter;
-
-@Entity
-@Table(name = "health_questions_reference")
-@Getter
-@Setter
-public class HealthQuestionReference {
-
-    @Id
-    private int questionNumber;
-
-    @Column(nullable = false, length = 500)
-    private String questionText;
-
-}
