--- src/main/java/com/example/Graduation/Project/entity/User.java
@@ -1,50 +1,0 @@
-package com.example.Graduation.Project.entity;
-
-import com.example.Graduation.Project.enums.UserRoleEnum;
-import jakarta.persistence.*;
-import lombok.Getter;
-import lombok.Setter;
-
-import java.time.LocalDateTime;
-
-@Entity
-@Table(
-        name = "user",
-        indexes = {
-                @Index(name = "uk_email", columnList = "email", unique = true)
-        }
-)
-@Getter
-@Setter
-public class User {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
-
-    @Column(nullable = false, unique = true)
-    private String email;
-
-    @Column(nullable = false)
-    private String password;
-
-    @Column(nullable = false)
-    private String name;
-
-    @Column(nullable = false)
-    private String surname;
-
-    @Enumerated(EnumType.STRING)
-    @Column(nullable = false)
-    private UserRoleEnum role;
-
-    @Column(columnDefinition = "int default 0")
-    private int memberCount;
-
-    @Column(columnDefinition = "timestamp default CURRENT_TIMESTAMP")
-    private LocalDateTime createdAt;
-
-    @Column(columnDefinition = "timestamp default CURRENT_TIMESTAMP")
-    private LocalDateTime updatedAt;
-
-}
