--- src/main/java/com/example/Graduation/Project/enums/MemberGenderEnum.java
@@ -1,6 +1,0 @@
-package com.example.Graduation.Project.enums;
-
-public enum MemberGenderEnum {
-    MALE,
-    FEMALE
-}
