--- src/main/java/com/example/Graduation/Project/enums/MemberPackagePaymentStatusEnum.java
@@ -1,7 +1,0 @@
-package com.example.Graduation.Project.enums;
-
-public enum MemberPackagePaymentStatusEnum {
-    PENDING,
-    COMPLETED,
-    FAILED
-}
