--- src/main/java/com/example/Graduation/Project/enums/ReservationStatusEnum.java
@@ -1,9 +1,0 @@
-package com.example.Graduation.Project.enums;
-
-public enum ReservationStatusEnum {
-    PENDING,
-    CONFIRMED,
-    CANCELLED,
-    COMPLETED,
-    NO_SHOW
-}
