--- src/main/java/com/example/Graduation/Project/enums/StatusEnum.java
@@ -1,14 +1,0 @@
-package com.example.Graduation.Project.enums;
-
-public enum StatusEnum {
-    ACTIVE,
-    PENDING_ID_CARD_VERIFICATION,
-    PENDING_PHOTO_VERIFICATION,
-    PENDING_HEALTH_FORM,
-    PENDING_HEALTH_REPORT,
-    PENDING_DOCTOR_APPROVAL,
-    REJECTED_ID_CARD,
-    REJECTED_PHOTO,
-    REJECTED_HEALTH_REPORT,
-    DISABLED
-}
