--- src/main/java/com/example/Graduation/Project/enums/UserRoleEnum.java
@@ -1,7 +1,0 @@
-package com.example.Graduation.Project.enums;
-
-public enum UserRoleEnum {
-    ADMIN,
-    DOCTOR,
-    COACH
-}
