--- src/main/java/com/example/Graduation/Project/repos/HealthInfoRepository.java
@@ -1,8 +1,0 @@
-package com.example.Graduation.Project.repos;
-
-import com.example.Graduation.Project.entity.HealthInfo;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface HealthInfoRepository extends JpaRepository<HealthInfo, Integer> {
-    HealthInfo findByMemberId(Integer memberId);
-}
