--- src/main/java/com/example/Graduation/Project/repos/HealthQuestionReferenceRepository.java
@@ -1,7 +1,0 @@
-package com.example.Graduation.Project.repos;
-
-import com.example.Graduation.Project.entity.HealthQuestionReference;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface HealthQuestionReferenceRepository extends JpaRepository<HealthQuestionReference, Integer> {
-}
