--- src/main/java/com/example/Graduation/Project/repos/MemberPackageRepository.java
@@ -1,7 +1,0 @@
-package com.example.Graduation.Project.repos;
-
-import com.example.Graduation.Project.entity.MemberPackage;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface MemberPackageRepository extends JpaRepository<MemberPackage, Integer> {
-}
