--- src/main/java/com/example/Graduation/Project/repos/PackageTypeRepository.java
@@ -1,7 +1,0 @@
-package com.example.Graduation.Project.repos;
-
-import com.example.Graduation.Project.entity.PackageType;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface PackageTypeRepository extends JpaRepository<PackageType, Integer> {
-}
