--- src/main/java/com/example/Graduation/Project/repos/PoolRepository.java
@@ -1,7 +1,0 @@
-package com.example.Graduation.Project.repos;
-
-import com.example.Graduation.Project.entity.Pool;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface PoolRepository extends JpaRepository<Pool, Integer> {
-}
