--- src/main/java/com/example/Graduation/Project/repos/SessionRepository.java
@@ -1,7 +1,0 @@
-package com.example.Graduation.Project.repos;
-
-import com.example.Graduation.Project.entity.Session;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface SessionRepository extends JpaRepository<Session, Integer> {
-}
