--- src/main/java/com/example/Graduation/Project/repos/UserRepository.java
@@ -1,9 +1,0 @@
-package com.example.Graduation.Project.repos;
-
-import com.example.Graduation.Project.entity.User;
-import org.springframework.data.jpa.repository.JpaRepository;
-import java.util.Optional;
-
-public interface UserRepository extends JpaRepository<User, Integer> {
-    Optional<User> findByEmail(String email);
-}
