+++ src/main/java/com/swimmingpool/SwimmingPool.java
@@ -1,0 +1,13 @@
+package com.swimmingpool;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class SwimmingPool {
+
+	public static void main(String[] args) {
+		SpringApplication.run(SwimmingPool.class, args);
+	}
+
+}
