+++ src/main/java/com/swimmingpool/model/entity/Session.java
@@ -1,0 +1,51 @@
+package com.swimmingpool.model.entity;
+
+import jakarta.persistence.*;
+import lombok.Getter;
+import lombok.Setter;
+
+import java.time.LocalDate;
+import java.time.LocalDateTime;
+import java.time.LocalTime;
+
+@Entity
+@Table(
+        name = "session",
+        indexes = {
+                @Index(name = "uk_pool_session", columnList = "poolId, sessionDate, startTime", unique = true),
+                @Index(name = "idx_session_date", columnList = "sessionDate")
+        }
+)
+@Getter
+@Setter
+public class Session {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private int id;
+
+    @Column(nullable = false)
+    private int poolId;
+
+    @Column(nullable = false)
+    private LocalDate sessionDate;
+
+    @Column(nullable = false)
+    private LocalTime startTime;
+
+    @Column(nullable = false)
+    private LocalTime endTime;
+
+    @Column(nullable = false)
+    private int capacity;
+
+    @Column(columnDefinition = "int default 0")
+    private int currentBookings;
+
+    @Column(nullable = false)
+    private boolean isEducationSession;
+
+    @Column(columnDefinition = "timestamp default CURRENT_TIMESTAMP")
+    private LocalDateTime createdAt;
+
+}
