+++ src/main/java/com/swimmingpool/model/enums/MemberGenderEnum.java
@@ -1,0 +1,6 @@
+package com.swimmingpool.model.enums;
+
+public enum MemberGenderEnum {
+    MALE,
+    FEMALE
+}
