+++ src/main/java/com/swimmingpool/model/enums/MemberPackagePaymentStatusEnum.java
@@ -1,0 +1,7 @@
+package com.swimmingpool.model.enums;
+
+public enum MemberPackagePaymentStatusEnum {
+    PENDING,
+    COMPLETED,
+    FAILED
+}
