+++ src/main/java/com/swimmingpool/model/enums/ReservationStatusEnum.java
@@ -1,0 +1,9 @@
+package com.swimmingpool.model.enums;
+
+public enum ReservationStatusEnum {
+    PENDING,
+    CONFIRMED,
+    CANCELLED,
+    COMPLETED,
+    NO_SHOW
+}
