+++ src/main/java/com/swimmingpool/model/enums/StatusEnum.java
@@ -1,0 +1,14 @@
+package com.swimmingpool.model.enums;
+
+public enum StatusEnum {
+    ACTIVE,
+    PENDING_ID_CARD_VERIFICATION,
+    PENDING_PHOTO_VERIFICATION,
+    PENDING_HEALTH_FORM,
+    PENDING_HEALTH_REPORT,
+    PENDING_DOCTOR_APPROVAL,
+    REJECTED_ID_CARD,
+    REJECTED_PHOTO,
+    REJECTED_HEALTH_REPORT,
+    DISABLED
+}
