+++ src/main/java/com/swimmingpool/model/enums/UserRoleEnum.java
@@ -1,0 +1,7 @@
+package com.swimmingpool.model.enums;
+
+public enum UserRoleEnum {
+    ADMIN,
+    DOCTOR,
+    COACH
+}
