+++ src/main/java/com/swimmingpool/repos/HealthInfoRepository.java
@@ -1,0 +1,8 @@
+package com.swimmingpool.repos;
+
+import com.swimmingpool.model.entity.HealthInfo;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface HealthInfoRepository extends JpaRepository<HealthInfo, Integer> {
+    HealthInfo findByMemberId(Integer memberId);
+}
