+++ src/main/java/com/swimmingpool/repos/HealthQuestionReferenceRepository.java
@@ -1,0 +1,7 @@
+package com.swimmingpool.repos;
+
+import com.swimmingpool.model.entity.HealthQuestionReference;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface HealthQuestionReferenceRepository extends JpaRepository<HealthQuestionReference, Integer> {
+}
