+++ src/main/java/com/swimmingpool/repos/MemberPackageRepository.java
@@ -1,0 +1,7 @@
+package com.swimmingpool.repos;
+
+import com.swimmingpool.model.entity.MemberPackage;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface MemberPackageRepository extends JpaRepository<MemberPackage, Integer> {
+}
