+++ src/main/java/com/swimmingpool/repos/PackageTypeRepository.java
@@ -1,0 +1,7 @@
+package com.swimmingpool.repos;
+
+import com.swimmingpool.model.entity.PackageType;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface PackageTypeRepository extends JpaRepository<PackageType, Integer> {
+}
