+++ src/main/java/com/swimmingpool/repos/ReservationRepository.java
@@ -1,0 +1,7 @@
+package com.swimmingpool.repos;
+
+import com.swimmingpool.model.entity.Reservation;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface ReservationRepository extends JpaRepository<Reservation, Integer> {
+}
