+++ src/main/java/com/swimmingpool/repos/SessionRepository.java
@@ -1,0 +1,7 @@
+package com.swimmingpool.repos;
+
+import com.swimmingpool.model.entity.Session;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface SessionRepository extends JpaRepository<Session, Integer> {
+}
