+++ src/main/java/com/swimmingpool/repos/UserRepository.java
@@ -1,0 +1,9 @@
+package com.swimmingpool.repos;
+
+import com.swimmingpool.model.entity.User;
+import org.springframework.data.jpa.repository.JpaRepository;
+import java.util.Optional;
+
+public interface UserRepository extends JpaRepository<User, Integer> {
+    Optional<User> findByEmail(String email);
+}
