+++ src/main/resources/application.properties
@@ -1,7 +1,7 @@
 spring.application.name=Graduation-Project
 server.port=8080
 
-spring.datasource.url=jdbc:mysql://cs401-402-swimmingpool-cs401-402.j.aivencloud.com:22282/defaultdb?verifyServerCertificate=true&useSSL=true&requireSSL=true&serverTimezone=UTC
+spring.datasource.url=jdbc:mysql://cs401-402-swimmingpool-cs401-402.j.aivencloud.com:22282/defaultdb
 spring.datasource.username=avnadmin
 spring.datasource.password=AVNS_oanhZgVS8KCHzU4fIKM
 spring.datasource.driver-class-name=com.mysql.cj.jdbc.Driver
