--- src/test/java/com/example/Graduation/Project/GraduationProjectApplicationTests.java
@@ -1,13 +1,0 @@
-package com.example.Graduation.Project;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class GraduationProjectApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
