+++ src/test/java/com/swimmingpool/GraduationProjectApplicationTests.java
@@ -1,0 +1,13 @@
+package com.swimmingpool;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class GraduationProjectApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
